requirements.txt
requests==2.32.3
web3==6.19.0
eth-account==0.11.3
eth-typing==5.0.1
eth-utils==5.1.0
loguru==0.7.2
fake-useragent==1.5.1
aiohttp-socks==0.9.0